UPDATE: внёс правки по ревью (модификаторы доступа, удалить @Component, переписать на стримы цикл и переделать консюмер

+++ stats/analyzer/src/main/java/analyzer/model/EventSimilarity.java
@@ -26,7 +26,7 @@ public class EventSimilarity {
     Long eventIdB;
 
     @Column(name = "result")
-    Float maxResult;
+    double maxResult;
 
     @Column(name = "time")
     LocalDateTime time;
